# Import http-date's utcTimeToRfc1123 function.