Another additional 35% reqs/second improvement.

Before:

```
Running 2m test @ http://localhost:3000/
  1 threads and 8 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   335.75us  221.81us  25.10ms   96.93%
    Req/Sec    23.27k     1.71k   28.00k    80.42%
  2778439 requests in 2.00m, 0.98GB read
Requests/sec:  23153.75
Transfer/sec:      8.38MB
```

After:

```
Running 2m test @ http://localhost:3000/
  1 threads and 8 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   267.73us  361.03us  39.96ms   99.19%
    Req/Sec    29.74k     3.99k   35.91k    63.17%
  3550827 requests in 2.00m, 1.26GB read
Requests/sec:  29590.07
Transfer/sec:     10.71MB
```