@reiddraper 2016-09-26
I believe the difference is in working w/ `UTCTime` vs `http-date`'s `HTTPDate` type.