@tmcgilchrist 2016-09-27
Curious if you've seen [thyme](https://hackage.haskell.org/package/thyme) which aims to provide a faster time implementation. 

My main concern is adding date/time parsing into airship directly, where it could live in it's own library. If the existing solutions don't work in this case (speed / data types), then I feel you're completely justified to build your own.