@reiddraper 2016-10-03
@tmcgilchrist `thyme` definitely looks neat. We have (so far) been happy using `UTCTime` everywhere, and though that that would be the best thing to support for Airship. However, definitely open to re-evaluating that.