@kellymclaughlin 2016-10-03
It seems that maybe `thyme` isn't being well supported at the moment. The build is marked as failing on the github repo and the last commit was back in April. Maybe the prudent move is to roll with this change for now and keep on eye on `thyme` to see if support picks up on it and try to get an RFC1123 helper submitted there if it does or make a switch to a library if we find one with suitable data types and performance.