@charleso 2016-10-05
👍  although an property for testing against the old/slow behaviour probably couldn't hurt