@patrickt 2016-10-05
I tested out a version with `thyme`; beyond incurring several extra dependencies, it is not as fast as the http-date version.