@tmcgilchrist 2016-10-05
Cool. Merge away :camel:

On Thursday, 6 October 2016, Patrick Thomson <notifications@github.com>
wrote:

> I tested out a version with thyme; beyond incurring several extra
> dependencies, it is not as fast as the http-date version.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/helium/airship/pull/109#issuecomment-251692424>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAKbuX_0zSBxD5kPLPmgyvlwC_sblBLsks5qw7UogaJpZM4KAmPo>
> .
>